+++ src/history_item.cpp
@@ -14,28 +14,10 @@ const std::string& HistoryItem::get_message() const {
     return message;
 }
 
-OpenChart::OpenChart(const better::Chart& c, const std::string& difficulty) : notes(c.notes) {
-    if (c.level) {
-        message = fmt::format(
-            "Opened Chart {} (level {})",
-            difficulty,
-            c.level->format("f")
-        );
-    } else {
-        message = fmt::format(
-            "Opened Chart {} (no level defined)",
-            difficulty
-        );
-    }
-}
-
-void OpenChart::do_action(EditorState& ed) const {
-    if (ed.chart_state) {
-        ed.chart_state->chart.notes = notes;
-    }
-}
-
-AddNotes::AddNotes(const better::Notes& notes) : notes(notes) {
+AddNotes::AddNotes(const std::string& chart, const better::Notes& notes) :
+    difficulty_name(chart),
+    notes(notes)
+{
     if (notes.empty()) {
         throw std::invalid_argument(
             "Can't construct a AddedNotes History Action with an empty note "
@@ -43,15 +25,19 @@ AddNotes::AddNotes(const better::Notes& notes) : notes(notes) {
         );
     }
     message = fmt::format(
-        "Added {} Note{}",
+        "Added {} note{} to chart {}",
         notes.size(),
-        notes.size() > 1 ? "s" : ""
+        notes.size() > 1 ? "s" : "",
+        chart
     );
 }
 
 void AddNotes::do_action(EditorState& ed) const {
     ed.set_playback_position(ed.time_at(notes.begin()->second.get_time()));
     if (ed.chart_state) {
+        if (not (ed.chart_state->difficulty_name == difficulty_name)) {
+            ed.open_chart(difficulty_name);
+        }
         for (const auto& [_, note] : notes) {
             ed.chart_state->chart.notes.insert(note);
         }
@@ -61,6 +47,9 @@ void AddNotes::do_action(EditorState& ed) const {
 void AddNotes::undo_action(EditorState& ed) const {
     ed.set_playback_position(ed.time_at(notes.begin()->second.get_time()));
     if (ed.chart_state) {
+        if (not (ed.chart_state->difficulty_name == difficulty_name)) {
+            ed.open_chart(difficulty_name);
+        }
         for (const auto& [_, note] : notes) {
             ed.chart_state->chart.notes.erase(note);
         }
@@ -68,7 +57,9 @@ void AddNotes::undo_action(EditorState& ed) const {
 }
 
 
-RemoveNotes::RemoveNotes(const better::Notes& notes) : AddNotes(notes) {
+RemoveNotes::RemoveNotes(const std::string& chart, const better::Notes& notes) :
+    AddNotes(chart, notes)
+{
     if (notes.empty()) {
         throw std::invalid_argument(
             "Can't construct a RemovedNotes History Action with an empty note "
@@ -76,9 +67,10 @@ RemoveNotes::RemoveNotes(const better::Notes& notes) : AddNotes(notes) {
         );
     }
     message = fmt::format(
-        "Removed {} Note{}",
+        "Removed {} note{} from chart {}",
         notes.size(),
-        notes.size() > 1 ? "s" : ""
+        notes.size() > 1 ? "s" : "",
+        chart
     );
 }
 
@@ -90,6 +82,81 @@ void RemoveNotes::undo_action(EditorState& ed) const {
     AddNotes::do_action(ed);
 }
 
-std::string get_message(const std::shared_ptr<HistoryItem>& awm) {
-    return awm->get_message();
+RerateChart::RerateChart(
+    const std::string& chart,
+    const std::optional<Decimal>& old_level,
+    const std::optional<Decimal>& new_level
+) :
+    chart(chart),
+    old_level(old_level),
+    new_level(new_level)
+{
+    message = fmt::format(
+        "Rerated {} : {} -> {}",
+        chart,
+        better::stringify_level(old_level),
+        better::stringify_level(new_level)
+    );
+}
+
+void RerateChart::do_action(EditorState& ed) const {
+    auto modified_chart = ed.song.charts.extract(chart);
+    modified_chart.mapped().level = new_level;
+    const auto [_1, inserted, _2] = ed.song.charts.insert(std::move(modified_chart));
+    if (not inserted) {
+        throw std::runtime_error(
+            "Redoing the change of chart level failed : "
+            "inserting the modified chart in the song object failed"
+        );
+    }
+}
+
+void RerateChart::undo_action(EditorState& ed) const {
+    auto modified_chart = ed.song.charts.extract(chart);
+    modified_chart.mapped().level = old_level;
+    const auto [_1, inserted, _2] = ed.song.charts.insert(std::move(modified_chart));
+    if (not inserted) {
+        throw std::runtime_error(
+            "Undoing the change of chart level failed : "
+            "inserting the modified chart in the song object failed"
+        );
+    }
+}
+
+RenameChart::RenameChart(
+    const std::string& old_name,
+    const std::string& new_name
+) :
+    old_name(old_name),
+    new_name(new_name)
+{
+    message = fmt::format(
+        "Rename {} -> {}",
+        old_name,
+        new_name
+    );
+}
+
+void RenameChart::do_action(EditorState& ed) const {
+    auto modified_chart = ed.song.charts.extract(old_name);
+    modified_chart.key() = new_name;
+    const auto [_1, inserted, _2] = ed.song.charts.insert(std::move(modified_chart));
+    if (not inserted) {
+        throw std::runtime_error(
+            "Redoing the renaming of the chart failed : "
+            "inserting the modified chart in the song object failed"
+        );
+    }
+}
+
+void RenameChart::undo_action(EditorState& ed) const {
+    auto modified_chart = ed.song.charts.extract(new_name);
+    modified_chart.key() = old_name;
+    const auto [_1, inserted, _2] = ed.song.charts.insert(std::move(modified_chart));
+    if (not inserted) {
+        throw std::runtime_error(
+            "Undoing the renaming of the chart failed : "
+            "inserting the modified chart in the song object failed"
+        );
+    }
 }